+++ src/main/java/com/yuuko/core/commands/core/commands/SettingsCommand.java
@@ -4,9 +4,7 @@ import com.yuuko.core.Cache;
 import com.yuuko.core.Configuration;
 import com.yuuko.core.commands.Command;
 import com.yuuko.core.commands.core.CoreModule;
-import com.yuuko.core.commands.core.settings.PrefixSetting;
-import com.yuuko.core.commands.core.settings.SettingExecuteBoolean;
-import com.yuuko.core.commands.core.settings.StarboardSetting;
+import com.yuuko.core.commands.core.settings.*;
 import com.yuuko.core.database.DatabaseFunctions;
 import com.yuuko.core.utilities.MessageHandler;
 import com.yuuko.core.utilities.Sanitiser;
@@ -22,10 +20,11 @@ public class SettingsCommand extends Command {
     private static final String[] settings = new String[]{
             "prefix",
             "deleteexecuted",
-            "commandlogging",
+            "commandlog",
             "nowplaying",
             "djmode",
-            "welcomemembers"
+            "newmember",
+            "starboard"
     };
 
     public SettingsCommand() {
@@ -39,46 +38,49 @@ public class SettingsCommand extends Command {
                 String[] commandParameters = command[1].split("\\s+", 2);
 
                 // Check to make sure the command is a valid command.
-                if(Arrays.asList(settings).contains(commandParameters[0].toLowerCase())) {
+                if(!Arrays.asList(settings).contains(commandParameters[0].toLowerCase())) {
                     EmbedBuilder embed = new EmbedBuilder().setTitle("_" + commandParameters[1].toUpperCase() + "_ is not a valid setting.");
                     MessageHandler.sendMessage(e, embed.build());
                     return;
                 }
 
                 if(!Sanitiser.checkParameters(e, command, 2)) {
+                    // We expect 0 in the super class to enable vision of all settings, but then check for 2 since that's the minimum used.
                     return;
                 }
 
-                if(commandParameters[0].equalsIgnoreCase("prefix")) {
-                    new PrefixSetting(e, commandParameters[1]);
-                }
-
-                if(commandParameters[0].equalsIgnoreCase("starboard")) {
-                    new StarboardSetting(e, commandParameters[1]);
-                } else {
-                    if(!commandParameters[1].equalsIgnoreCase("true") && !commandParameters[1].equalsIgnoreCase("false")) {
-                        EmbedBuilder embed = new EmbedBuilder().setTitle("_" + commandParameters[1].toUpperCase() + "_ is not a valid value. (Valid: TRUE, FALSE)");
-                        MessageHandler.sendMessage(e, embed.build());
+                switch(commandParameters[0].toLowerCase()) {
+                    case "prefix": new PrefixSetting(e, commandParameters[1]);
                         return;
-                    }
-                    new SettingExecuteBoolean(e, commandParameters[0].toLowerCase(), commandParameters[1]);
+                    case "starboard": new StarboardSetting(e, commandParameters[1]);
+                        return;
+                    case "commandlog": new CommandLogSetting(e, commandParameters[1]);
+                        return;
+                    case "newmember": new NewMemberSetting(e);
+                        return;
+                    default:
+                        if(!commandParameters[1].equalsIgnoreCase("true") && !commandParameters[1].equalsIgnoreCase("false")) {
+                            EmbedBuilder embed = new EmbedBuilder().setTitle("_" + commandParameters[1].toUpperCase() + "_ is not a valid value. (Valid: TRUE, FALSE)");
+                            MessageHandler.sendMessage(e, embed.build());
+                            return;
+                        }
+                        new SettingExecuteBoolean(e, commandParameters[0].toLowerCase(), commandParameters[1]);
                 }
 
             } else {
-                ArrayList<Boolean> settings = DatabaseFunctions.getGuildSettings(e.getGuild().getId());
-                String starboard = DatabaseFunctions.getGuildSetting("starboard", e.getGuild().getId());
+                ArrayList<String> settings = DatabaseFunctions.getGuildSettings(e.getGuild().getId());
 
                 // Embed displaying all of the current settings for the server, giving information about each setting.
                 EmbedBuilder commandModules = new EmbedBuilder()
                         .setTitle("Settings for **" + e.getGuild().getName() + "**")
-                        .setDescription("Settings can be changed by typing '<prefix>settings [setting] [value]' where [setting] is a value found below and [value] is a valid value, with special values like booleans being either TRUE or FALSE (case insensitive)")
-                            .addField("prefix", "**" + DatabaseFunctions.getGuildSetting("commandPrefix", e.getGuild().getId()) + "** - The message prefix used to symbolise a command.", false)
-                            .addField("deleteExecuted", "**" + settings.get(0) + "** - Deletes the users command string when it is executed.", false)
-                            .addField("commandLogging", "**" + settings.get(1) + "** - Sends executed commands to a predefined logging channel.", false)
-                            .addField("nowPlaying", "**" + settings.get(2) + "** - Sends information of the current track when it changes.", false)
-                            .addField("djMode", "**" + settings.get(3) + "** - Defines if DJ mode is on, meaning only users with the role 'DJ' can use certain audio commands.", false)
-                            .addField("welcomeMembers", "**" + settings.get(4) + "** - Whether or not Yuuko will greet each new member that joins the server.", false)
-                            .addField("starboard", "**" + (!starboard.equals("") ? e.getGuild().getTextChannelById(starboard).getName() : "None") + "** - The set starboard channel, where any messages reacted to with a star will be sent.", false)
+                        .setDescription("Settings can be changed by typing **"+settings.get(0)+"settings [setting] [value]** where **[setting]** is a value found below and **[value]** is a valid value, with special values like booleans being either **TRUE** or **FALSE** (case insensitive)")
+                            .addField("prefix (String)", "**__" + settings.get(0) + "__** - The message prefix used to symbolise a command.", false)
+                            .addField("deleteExecuted (Boolean)", "**__" + settings.get(1) + "__** - Deletes the users command string when it is executed.", false)
+                            .addField("nowPlaying (Boolean)", "**__" + settings.get(2) + "__** - Sends information of the current track when it changes.", false)
+                            .addField("djMode (Boolean)", "**__" + settings.get(3) + "__** - Defines if DJ mode is on, meaning only users with the role 'DJ' can use certain audio commands.", false)
+                            .addField("newMember (Mention)", (settings.get(4) != null ? e.getGuild().getTextChannelById(settings.get(4)).getAsMention() : "**__Disabled__**") + " - Where Yuuko will greet each new member that joins the server.", false)
+                            .addField("starboard (Mention)", (settings.get(5) != null ? e.getGuild().getTextChannelById(settings.get(5)).getAsMention() : "**__Disabled__**") + " - Where any messages reacted to with a ⭐ will be sent.", false)
+                            .addField("commandLog (Mention)", (settings.get(6) != null ? e.getGuild().getTextChannelById(settings.get(6)).getAsMention() : "**__Disabled__**") + " - Sends executed commands to a defined log channel.", false)
                         .setFooter(Cache.STANDARD_STRINGS[1] + e.getMember().getEffectiveName(), e.getGuild().getMemberById(Configuration.BOT_ID).getUser().getAvatarUrl());
                 MessageHandler.sendMessage(e, commandModules.build());
             }