+++ src/main/java/com/yuuko/core/commands/core/settings/CommandLogSetting.java
@@ -1,0 +1,66 @@
+package com.yuuko.core.commands.core.settings;
+
+import com.yuuko.core.database.DatabaseFunctions;
+import com.yuuko.core.utilities.MessageHandler;
+import com.yuuko.core.utilities.MessageUtility;
+import net.dv8tion.jda.core.EmbedBuilder;
+import net.dv8tion.jda.core.Permission;
+import net.dv8tion.jda.core.entities.TextChannel;
+import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
+
+public class CommandLogSetting {
+
+    public CommandLogSetting(MessageReceivedEvent e, String value) {
+        executeCommand(e, value);
+    }
+
+    public void executeCommand(MessageReceivedEvent e, String value) {
+        if(value.equalsIgnoreCase("setup")) {
+            if(e.getGuild().getSelfMember().hasPermission(Permission.MANAGE_CHANNEL, Permission.MANAGE_PERMISSIONS)) {
+                setupCommandLog(e);
+            } else {
+                EmbedBuilder embed = new EmbedBuilder().setTitle("Missing Permission").setDescription("I require the **Manage Channel** and **Manage Permissions** permissions to setup the command log automatically.");
+                MessageHandler.sendMessage(e, embed.build());
+            }
+            return;
+        }
+
+        TextChannel channel = MessageUtility.getFirstMentionedChannel(e);
+        if(channel != null) {
+            if(DatabaseFunctions.setGuildSettings("commandLog", channel.getId(), e.getGuild().getId())) {
+                EmbedBuilder embed = new EmbedBuilder().setTitle("Command Log").setDescription("The command log has been set to " + channel.getAsMention() + ".");
+                MessageHandler.sendMessage(e, embed.build());
+            }
+        } else {
+            if(DatabaseFunctions.setGuildSettings("commandLog", null, e.getGuild().getId())) {
+                EmbedBuilder embed = new EmbedBuilder().setTitle("Command Log").setDescription("The command log has been unset, thus deactivating the command log.");
+                MessageHandler.sendMessage(e, embed.build());
+            }
+        }
+    }
+
+    private void setupCommandLog(MessageReceivedEvent e) {
+        try {
+            e.getGuild().getController().createTextChannel("command-log").queue(channel -> {
+                TextChannel textChannel = (TextChannel)channel;
+                channel.createPermissionOverride(e.getGuild().getSelfMember()).setAllow(Permission.MESSAGE_WRITE).queue();
+                if(DatabaseFunctions.setGuildSettings("commandLog", channel.getId(), e.getGuild().getId())) {
+                    EmbedBuilder embed = new EmbedBuilder().setTitle("Command Log").setDescription("The " + textChannel.getAsMention() + " channel has been setup correctly.");
+                    MessageHandler.sendMessage(e, embed.build());
+                }
+            });
+        } catch(Exception ex) {
+            //
+        }
+    }
+
+    /**
+     * Executes the logging feature of the bot.
+     * @param e MessageReceivedEvent
+     * @param executionTimeMs long
+     */
+    public static void executeLogging(MessageReceivedEvent e, long executionTimeMs) {
+        TextChannel log = e.getGuild().getTextChannelById(DatabaseFunctions.getGuildSetting("commandLog", e.getGuild().getId()));
+        MessageHandler.sendMessage(log, "```" + e.getAuthor().getName() + "#" + e.getAuthor().getDiscriminator() + " used command " + e.getMessage().getContentDisplay() + " in " + e.getMessage().getChannel().getName() + ". (Execution time: "+ executionTimeMs +"ms)```");
+    }
+}