+++ src/main/java/com/yuuko/core/commands/core/settings/SettingExecuteBoolean.java
@@ -1,14 +1,11 @@
 package com.yuuko.core.commands.core.settings;
 
-import com.yuuko.core.Configuration;
 import com.yuuko.core.database.DatabaseFunctions;
 import com.yuuko.core.utilities.MessageHandler;
 import net.dv8tion.jda.core.EmbedBuilder;
-import net.dv8tion.jda.core.entities.TextChannel;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
 import java.awt.*;
-import java.util.List;
 
 public class SettingExecuteBoolean {
 
@@ -33,20 +30,6 @@ public class SettingExecuteBoolean {
             }
         } catch(Exception ex) {
             MessageHandler.sendException(ex, setting + "[" + value + "]");
-        }
-    }
-
-    /**
-     * Executes the logging feature of the bot.
-     * @param e MessageReceivedEvent
-     * @param executionTimeMs long
-     */
-    public void executeLogging(MessageReceivedEvent e, long executionTimeMs) {
-        try {
-            List<TextChannel> log = e.getGuild().getTextChannelsByName("command-log", true);
-            log.get(0).sendMessage("```" + e.getAuthor().getName() + "#" + e.getAuthor().getDiscriminator() + " used command " + e.getMessage().getContentDisplay() + " in " + e.getMessage().getChannel().getName() + ". (Execution time: "+ executionTimeMs +"ms)```").queue();
-        } catch(Exception ex) {
-            MessageHandler.sendMessage(e, "This server has logging enabled but has no 'command-log' text channel, please add that channel or disable the module by typing the '" + Configuration.GLOBAL_PREFIX + "module logging' command to stop this message.");
         }
     }